pyproject.toml
[project]
name = "floor-plan-from-image"
version = "0.1.0"
description = ""
authors = [
    {name = "Vladimir",email = "vpd63@inbox.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "open3d (>=0.19.0,<0.20.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "ultralytics (>=8.3.228,<9.0.0)",
    "pycolmap (>=3.13.0,<4.0.0)",
    "trimesh (>=4.9.0,<5.0.0)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
